requirements.txt
aiofiles==0.8.0
discord_py_interactions==4.3.1
py_expression_eval==0.3.14
pyfiglet==0.8.post1
python-dotenv==0.20.0
